+++ tests/test_lint.py
@@ -52,7 +52,7 @@ def test_lint_git(git_repository):
 def test_lint_submodule(submodule_repository):
     """Extremely simple test for lint with an ignored submodule."""
     project = Project(submodule_repository)
-    (submodule_repository / "submodule/foo.c").touch()
+    (submodule_repository / "submodule/foo.c").write_text("foo")
     report = ProjectReport.generate(project)
     result = lint(report)
     assert result
@@ -61,7 +61,7 @@ def test_lint_submodule(submodule_repository):
 def test_lint_submodule_included(submodule_repository):
     """Extremely simple test for lint with an included submodule."""
     project = Project(submodule_repository, include_submodules=True)
-    (submodule_repository / "submodule/foo.c").touch()
+    (submodule_repository / "submodule/foo.c").write_text("foo")
     report = ProjectReport.generate(project)
     result = lint(report)
     assert not result
@@ -124,7 +124,7 @@ def test_lint_missing_licenses(fake_repository, stringio):
 
 def test_lint_unused_licenses(fake_repository, stringio):
     """An unused license is detected."""
-    (fake_repository / "LICENSES/MIT.txt").touch()
+    (fake_repository / "LICENSES/MIT.txt").write_text("foo")
     project = Project(fake_repository)
     report = ProjectReport.generate(project)
     lint_summary(report, out=stringio)
@@ -136,7 +136,7 @@ def test_lint_unused_licenses(fake_repository, stringio):
 @posix
 def test_lint_read_errors(fake_repository, stringio):
     """A read error is detected."""
-    (fake_repository / "foo.py").touch()
+    (fake_repository / "foo.py").write_text("foo")
     (fake_repository / "foo.py").chmod(0o000)
     project = Project(fake_repository)
     report = ProjectReport.generate(project)
@@ -148,7 +148,7 @@ def test_lint_read_errors(fake_repository, stringio):
 
 def test_lint_files_without_copyright_and_licensing(fake_repository, stringio):
     """A file without copyright and licensing is detected."""
-    (fake_repository / "foo.py").touch()
+    (fake_repository / "foo.py").write_text("foo")
     project = Project(fake_repository)
     report = ProjectReport.generate(project)
     result = lint_files_without_copyright_and_licensing(report, out=stringio)